atualização apos a aula 6

+++ tests/test_app.py
@@ -80,17 +80,18 @@ def test_create_email_integrity_error(client):
     assert response.json() == {"detail": "Email already exists"}
 
 
-def test_read_users_with_users(client, user):
+def test_read_users(client, user, token):
     user_schema = UserPublic.model_validate(user).model_dump()
-    response = client.get("/users/")
+    response = client.get("/users/", headers={"Authorization": f"Bearer {token}"})
 
     assert response.status_code == HTTPStatus.OK
     assert response.json() == {"users": [user_schema]}
 
 
-def test_update_user_ok(client, user):
+def test_update_user_ok(client, user, token):
     response = client.put(
         "/users/1",
+        headers={"Authorization": f"Bearer {token}"},
         json={
             "username": "bob",
             "email": "bob@example.com",
@@ -105,29 +106,32 @@ def test_update_user_ok(client, user):
     }
 
 
-def test_update_user_return_not_found(client, user):
+def test_update_user_return_not_found(client, user, token):
     response = client.put(
-        "/users/-2",
+        "/users/-666",
+        headers={"Authorization": f"Bearer {token}"},
         json={
             "username": "bob",
             "email": "bob@example.com",
             "password": "123456",
         },
     )
-    assert response.status_code == HTTPStatus.NOT_FOUND
-    assert response.json() == {"detail": "User not found"}
+    assert response.status_code == HTTPStatus.FORBIDDEN
+    assert response.json() == {"detail": "Not enough permission"}
 
 
-def test_delete_user_ok(client, user):
-    response = client.delete("/users/1")
+def test_delete_user_ok(client, user, token):
+    response = client.delete(
+        f"/users/{user.id}", headers={"Authorization": f"Bearer {token}"}
+    )
     assert response.status_code == HTTPStatus.OK
     assert response.json() == {"message": "User deleted"}
 
 
-def test_delete_user_return_not_found(client, user):
-    response = client.delete("/users/-2")
-    assert response.status_code == HTTPStatus.NOT_FOUND
-    assert response.json() == {"detail": "User not found"}
+def test_delete_user_return_not_found(client, user, token):
+    response = client.delete("/users/-2", headers={"Authorization": f"Bearer {token}"})
+    assert response.status_code == HTTPStatus.FORBIDDEN
+    assert response.json() == {"detail": "Not enough permission"}
 
 
 def test_get_user_ok(client, user):
@@ -143,7 +147,7 @@ def test_get_user_error(client, user):
     assert response.json() == {"detail": "User not found"}
 
 
-def test_update_integrity_error(client, user):
+def test_update_integrity_error(client, user, token):
     client.post(
         "/users",
         json={
@@ -154,6 +158,7 @@ def test_update_integrity_error(client, user):
     )
     response_update = client.put(
         f"/users/{user.id}",
+        headers={"Authorization": f"Bearer {token}"},
         json={
             "username": "fausto",
             "email": "bob@example.com",
@@ -172,3 +177,15 @@ def test_get_session():
     assert isinstance(session, Session)
 
     response.close()
+
+
+def test_get_token(client, user):
+    response = client.post(
+        "/token", data={"username": user.email, "password": user.clean_password}
+    )
+
+    token = response.json()
+
+    assert response.status_code == HTTPStatus.OK
+    assert token["token_type"] == "Bearer"
+    assert "access_token" in token
